requirements.txt
langchain-community
langchain-openai
langchain
pymupdf
flask
pypdf
python-dotenv
